requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
asyncpg==0.29.0
redis==5.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.23.0
stripe==9.4.0
loguru==0.7.2
alembic==1.13.1
psycopg2-binary==2.9.9 